Fix off-by-one in ioctl_EVIOCG_bits
This bug caused values at the end of the list to not be reported back.

Example: On my system, SW_PEN_INSERTED (0x0f) is the last switch constant. That is,
SW_MAX=SW_PEN_INSERTED. Other tools (python-libevdev) correctly reported this switch
as 'set' but python-evdev did not. This commit fixes this issue.